+++ src/shared/lib/context/ModalContext/ModalContext.ts
@@ -1,0 +1,14 @@
+import { createContext } from "react";
+import type { ReactNode } from "react";
+
+export interface ModalContextType {
+    isModalOpen: boolean;
+    openModal: () => void;
+    closeModal: () => void;
+}
+
+export const ModalContext = createContext<ModalContextType | undefined>(undefined);
+
+export interface ModalProviderProps {
+    children: ReactNode;
+}
