+++ src/shared/lib/context/ModalContext/index.ts
@@ -1,0 +1,3 @@
+export { ModalProvider } from "./ModalProvider";
+export { useModal } from "./useModal";
+export type { ModalContextType, ModalProviderProps } from "./ModalContext";
