+++ src/shared/lib/context/ModalContext/useModal.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { ModalContext } from "./ModalContext";
+
+export const useModal = () => {
+    const context = useContext(ModalContext);
+    if (context === undefined) {
+        throw new Error("useModal must be used within a ModalProvider");
+    }
+    return context;
+};
